% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_checks.R
\name{abc_crit}
\alias{abc_crit}
\title{ABC criterion for choosing the number of static factors}
\usage{
abc_crit(X, kmax, nbck = NULL, cmax = 3)
}
\arguments{
\item{X}{\eqn{T x N} data matrix}

\item{kmax}{maximum number of factors}

\item{nbck}{number of sublocks to be used, defaults to \code{floor(N/10)}}

\item{cmax}{maximum value for the penalty constant, default is 3}
}
\value{
\item{rhat1}{determines the number of shocks using a large window}
\item{rhat2}{determines the number of shocks using a small window}
}
\description{
The function estimates the number of static factors in the
dynamic factor model. These information criteria are based on
Alessi, Barigozzi and Capasso (2010). This function is a direct
translation of the MATLAB code given in http://www.barigozzi.eu/Codes.html
written by Matteo Barigozzi.
}
\examples{
(abc1 <- abc_crit(X = FRED_heavy$df, kmax = 25))
(abc2 <- abc_crit(X = FRED_light$df, kmax = 25))
}
\seealso{
Alessi, L., M. Barigozzi, and M. Capasso (2010).
Improved penalization for determining the number of factors
in approximate static factor models.
Statistics and Probability Letters 80, 1806-1813.
}
