% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_checks.R
\name{admissible_mods}
\alias{admissible_mods}
\title{Admissible models}
\usage{
admissible_mods(q, max_nu, fix_r = TRUE)
}
\arguments{
\item{q}{input dimension, number of shocks}

\item{max_nu}{maximum Kronecker index value}

\item{fix_r}{boolean, whether the state dimension of the state space
representation is fixed to \code{q*max_nu}}
}
\value{
a list of two lists of equal length, where each element in the list
correspond either to the Kronecker index specification or to the additional
possible restrictions imposed on the lag orders of \eqn{c(z)} and \eqn{d(z)}
for given DFM specification. The elements are saved in named lists
\item{nus}{list of Kronecker index vectors}
\item{degs}{list of polynomial degrees}
}
\description{
\code{admissible_mods} gives the qualified models according to
the model selection criteria in section 3.3 of \emph{Estimation of
Impulse-Response Functions with Dynamic Factor Models: A New Parametrization}
available at \url{https://arxiv.org/pdf/2202.00310.pdf}.
}
\note{
This is a rather crude way of obtaining the set of models satisfying the
model selection criteria because the function simulates
the model using parameter values from a \emph{U(-1,1)}-distribution and then
running the checks. This entails that the function may not always find the
maximum number of qualified models, and this being the case especially with
specifications where the input dimension and maximum Kronecker index are given
a high value. However, for a small values of these parameters the function
should work fine.
}
\examples{
tmp <- admissible_mods(4, 4)
sample(1:length(tmp), 1) \%>\% sapply(function(x)
  paste("The qualified model no. ", x,
  "has a Kronecker index structure ", tmp$nus[[x]], "with the
  orders of the lag polynomials restricted to ", tmp$degs[[x]]))
}
