% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_checks.R
\name{baingcriterion}
\alias{baingcriterion}
\title{Bai & Ng information criteria for determining the number of static factors}
\usage{
baingcriterion(X, rmax, rvar = NULL)
}
\arguments{
\item{X}{\eqn{T x N} data matrix}

\item{rmax}{maximum number of static factors to be considered}

\item{rvar}{fix penalty function scaling term, defaults to \code{rvar=rmax}}
}
\value{
lists of IC and PC, where the values indicate the suggested number of factors
}
\description{
This function calculates two different information criteria for
determining the number of static factors, which were introduced in
Bai & Ng (2002). The function gives two different sets of
information criteria (IC and PC), which use three different
penalty functions to take into account the increased complexity of the
model.
}
\examples{
(bn1 <- baingcriterion(X = FRED_heavy$df, 25))
(bn2 <- baingcriterion(X = FRED_light$df, 25))
}
\seealso{
Bai, J., & Ng, S. (2002). Determining the number of factors in approximate factor models.
Econometrica, 70(1), 191-221.
}
