% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{boot_irfs}
\alias{boot_irfs}
\title{Bootstrap confidence intervals for IRFs}
\usage{
boot_irfs(
  df,
  nu,
  degs,
  maxit,
  conv_crit,
  init0,
  verbose,
  h,
  nrep,
  int_vars,
  save_file = FALSE
)
}
\arguments{
\item{df}{\code{(T x N)}-dimensional data matrix}

\item{nu}{\code{q}-dimensional vector of Kronecker indices}

\item{degs}{optional, vector of length 2, defining the degrees of c(z) and d(z)}

\item{maxit}{Maximum number of iterations in the estimation procedure, defaults to 250.}

\item{conv_crit}{Convergence criterion in the estimation, defaults to 1e-3}

\item{init0}{specify either a non-negative integer with zero indicating no bootstrapping of initial values
and non-zero value defining the number of bootstrap draws or a list containing initial values of the SSM
and \eqn{(q x q)} covariance matrix of error term in state eqn.}

\item{verbose}{Should the estimation process be printed?}

\item{h}{number of periods for which IRFs are calculated}

\item{nrep}{number of bootstrap draws}

\item{save_file}{save IRF array at the end?}
}
\value{

}
\description{
Bootstrap confidence intervals for IRFs
}
