% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{cum_irf}
\alias{cum_irf}
\title{Cumulate IRFs for variables in differences}
\usage{
cum_irf(irf_array, trans_ix)
}
\arguments{
\item{irf_array}{\eqn{(N x q x (h+1))} array containing the impulse response function, with \eqn{N} denoting
the number of variables, \eqn{q} the number of shocks, \eqn{h} the IRF horizon}

\item{trans_ix}{vector of transformation codes according in McCracken & Ng (JME, 2016) format}
}
\value{
\eqn{(N x q x (h+1))} array containing the cumulated IRFs
}
\description{
This function cumulates IRFs such that variables in first and second differences
are cumulated once and twice, respectively.
}
\keyword{internal}
