% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_everything.R
\name{do_everything_svar}
\alias{do_everything_svar}
\title{Estimate IRFs and associated bootstrap intervals for SVAR}
\usage{
do_everything_svar(df, p, nrep = 500, h = 50, ci = 0.8)
}
\arguments{
\item{df}{list containing items:
\enumerate{
\item \code{df}, data in a tall matrix;
\item \code{int_ix}, index coding the rows of the interest variables in the data;
\item \code{trans_ix}, index giving the variable transformation codes
}}

\item{p}{VAR degree}

\item{nrep}{number of replications in the bootstrap}

\item{h}{h-step ahead IRFs}

\item{ci}{confidence intervals for the IRFs}
}
\value{

}
\description{
Estimate IRFs and associated bootstrap intervals for SVAR
}
