% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{estim_wrap}
\alias{estim_wrap}
\title{Estimation wrapper for RMFD-DF model}
\usage{
estim_wrap(
  df,
  nu,
  degs = NULL,
  maxit = 250,
  conv_crit = 0.001,
  init0 = 0,
  verbose = TRUE,
  h = 50
)
}
\arguments{
\item{df}{\code{(T x N)}-dimensional data matrix}

\item{nu}{\code{q}-dimensional vector of Kronecker indices}

\item{degs}{optional, vector of length 2, defining the degrees of c(z) and d(z)}

\item{maxit}{Maximum number of iterations in the estimation procedure, defaults to 250.}

\item{conv_crit}{Convergence criterion in the estimation, defaults to 1e-3}

\item{init0}{specify either a non-negative integer with zero indicating no bootstrapping of initial values
and non-zero value defining the number of bootstrap draws or a list containing initial values of the SSM
and \eqn{(q x q)} covariance matrix of error term in state eqn.}

\item{verbose}{Should the estimation process be printed?}
}
\value{
The function returns a list containing the final model in a state space
and RMFD format and the smoothed residuals series of the final model
}
\description{
This function estimates the model for a given data matrix and
a vector of Kronecker indices. The user must also specify the
maximum number of iterations used for the estimation as well as
the convergence criterion
}
\examples{
df <- rnorm(100)
}
