% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FG_data}
\alias{FG_data}
\title{High-dimensional macroeconomic data set used in Forni and Gambetti (2010)}
\format{
A list with the following elements
\describe{
\item{df}{\eqn{T x n} data frame where \eqn{T} is the time-series dimension,
and \eqn{n} is the cross-sectional dimension}
\item{data}{\code{Date} variable giving the time span of the data}
\item{trans_ix}{Variable transformation codes with the interpretation as in
McCracken and Ng (2016)}
}
}
\source{
\url{http://pareto.uab.es/lgambetti/ReplicaForniGambettiJME.zip}
}
\usage{
FG_data
}
\description{
This data is used by Forni and Gambetti (2010) in their monetary policy
study using dynamic factor models. The data spans the period between
March 1973 and November 2007 and includes 112 macroeconomic variables.
The data is included in this package for purposes of comparison to the
main data set used in the empirical exercise, FRED-MD. The method of obtaining
this data set from the raw data is described in the script \code{data-raw/DATASET.R}.
}
\references{
Forni, M., & Gambetti, L. (2010). The dynamic effects of monetary policy:
A structural factor model approach. Journal of Monetary Economics, 57(2), 203-216.

McCracken, M. W., & Ng, S. (2016). FRED-MD: A monthly database for macroeconomic research.
Journal of Business & Economic Statistics, 34(4), 574-589.
}
\keyword{datasets}
