% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{finalize_irf}
\alias{finalize_irf}
\title{NO EXPORT:: Finalize IRF for the plots}
\usage{
finalize_irf(irf_obj, shock_size, norm_id, trans_ix, int_vars)
}
\arguments{
\item{irf_obj}{array containing the structural impulse responses}

\item{shock_size}{size of the shock}

\item{norm_id}{vector of length two, which code the variable corresponding
to the shock of interest and its ordering in the system}

\item{trans_ix}{vector containing the transformation codes}

\item{int_vars}{vector coding the interest variables}
}
\value{
A matrix containing the IRFs as columns
}
\description{
This function transforms the structural impulse responses such
that they reflect the responses of the endogenous variables
to an exogenous shock of custom size, determined by the user.
Additionally, it accumulates impulse responses of differenced
variables such that they reflect the responses in levels.
Finally, variables in logs, log-differences or in relative differences
are multiplied by 100 such that they are transformed to percentages.
This function is intended for a single shock identification,
i.e. the output is a matrix with the columns containing the responses
of the variables to the shock.
}
\examples{
rnorm(100)
}
