% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRED_heavy}
\alias{FRED_heavy}
\title{FRED-MD (December 2021 vintage) data set with heavy transformations}
\format{
A list with the following elements
\describe{
\item{discarded}{data frame listing the discarded variables and the number of
missing values under the time span under consideration}
\item{df}{\eqn{T x n} data frame where \eqn{T} is the time-series dimension,
and \eqn{n} is the cross-sectional dimension}
\item{data}{\code{Date} variable giving the time span of the data}
\item{trans_ix}{Variable transformation codes with the interpretation as in
McCracken and Ng (2016)}
}
}
\source{
\url{https://files.stlouisfed.org/files/htdocs/fred-md/monthly/2021-12.csv}
}
\usage{
FRED_heavy
}
\description{
This data is taken from FRED-MD database, with the sample span corresponding to
that of \code{FG_data}, i.e. March 1973 to November 2007. The database consists
of 127 variables, out of which two are discarded due to many missing variables, and
the corresponding variable names can be accessed by \code{FRED_heavy$discarded}.
The data is transformed stationary according to the methodology given in
McCracken and Ng. The method of obtaining this data set from the raw data
is described in the script \code{data-raw/DATASET.R}.
}
\references{
McCracken, M. W., & Ng, S. (2016). FRED-MD: A monthly database for macroeconomic research.
Journal of Business & Economic Statistics, 34(4), 574-589.
}
\keyword{datasets}
