% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{get_acf_qnt}
\alias{get_acf_qnt}
\title{Get sample quantiles of the autocorrelation coefficients of the variables}
\usage{
get_acf_qnt(df, lag.acf = 8, qnt_grid = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{df}{\eqn{T x N} data matrix}

\item{lag.acf}{maximum number of lags at which the autocorrelations are calculated}

\item{qnt_grid}{vector of probabilities determining the sample percent quantiles}
}
\value{
Quantile table giving the distribution of autocorrelation coefficients at different lags
}
\description{
This function prints the percent quantiles of the autocorrelation coefficients
of the variables included in the data set. This is a simple way to
summarize the degree of non-stationarity and persistence of the
variables. The number of lags at which autocorrelations are calculated
and the grid of probabilities for the percent quantiles can be defined.
}
