% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intial_estimation.R
\name{init_vals}
\alias{init_vals}
\title{Initial values for RMFD model}
\usage{
init_vals(irf_obj, nu, degs = NULL, tmpl_mod)
}
\arguments{
\item{irf_obj}{impulse response function object of dimension (\eqn{n, q,} \code{lag.max}+1)}

\item{nu}{vector of  Kronecker indices}

\item{degs}{vector of length 2, where one can optionally specify the degrees of c(z) and d(z),
either of the parameters must be equal to \code{max(nu)}}

\item{tmpl_mod}{model template for extracting the initial values}
}
\value{
a list of elements
\item{init_val}{a vector of initial values}
\item{sigma}{error covariance matrix}
}
\description{
\code{init_vals} gives the initial values of the RMFD model in state space format.
The initial values are calculated from the input IRF that is used to generate the Hankel matrix,
which are used in the regressions to determine parameter matrices associated with the lag
polynomials c(z) and d(z).
}
\keyword{internal}
