% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intial_estimation.R
\name{init_wrap}
\alias{init_wrap}
\title{Obtain initial values corresponding to a given model specification}
\usage{
init_wrap(data, nu, degs = NULL, tmpl_mod, reg_prm = NULL)
}
\arguments{
\item{data}{matrix of dimension \eqn{(T x N)}, where\eqn{T} is the number of observations and
\eqn{N} is the number of variables}

\item{nu}{vector of right Kronecker indices}

\item{degs}{vector of length 2, where one can optionally specify the degrees of c(z) and d(z), either of the parameters must be equal to max(nu)}

\item{tmpl_mod}{model template, see \code{rmfd4dfm:::tmpl_rmfd_echelon_ss}}

\item{reg_prm}{regularization parameter governing the std deviation of the randomly drawn noise vector,
see also \link[rmfd4dfm]{estim_wrap}}
}
\value{
a list of elements
\item{params0}{a vector of initial parameters}
\item{sigma}{\code{q x q} initial error covariance matrix}
}
\description{
\code{init_wrap} estimates the initial values for a given model specification, this
function is a helper to the function \code{boot_init}
}
\keyword{internal}
