% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{irf_x}
\alias{irf_x}
\title{NO EXPORT:: Left and right multiply IRFs with constant matrices}
\usage{
irf_x(irf, pre_mat = NULL, post_mat = NULL)
}
\arguments{
\item{irf}{\code{N x q x h} array containing the impulse responses}

\item{pre_mat}{matrix to be left multiplied}

\item{post_mat}{matrix to be right multiplied}
}
\value{
The function \code{irf_x} returns the transformed IRF as an array
}
\description{
NO EXPORT:: Left and right multiply IRFs with constant matrices
}
\examples{
irf <- rnorm(100)
}
