% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfks_cpp}
\alias{kfks_cpp}
\title{Kalman filter smoother}
\usage{
kfks_cpp(A, C, R, Q, data_wide, P1, dim_state, only_ll)
}
\arguments{
\item{A}{n x r observation system matrix}

\item{C}{r x r state transition matrix}

\item{R}{n x n noise covariance matrix}

\item{Q}{q x q error covariance matrix}

\item{data_wide}{n x T data matrix}

\item{P1}{r x r initial state variance matrix}

\item{dim_state}{value r}

\item{only_ll}{boolean, return only log-likelihood calculated with KFss}
}
\description{
Kalman filter smoother
}
\keyword{internal}
