% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intial_estimation.R
\name{lr_approx}
\alias{lr_approx}
\title{Get Tall IRFs}
\usage{
lr_approx(ssmod, dim_in, lag.max = 30)
}
\arguments{
\item{ssmod}{square state space model object}

\item{dim_in}{the dynamic factor dimension of the DFM}

\item{lag.max}{number of lags, default is 30}
}
\value{
a list of elements:
\item{irf}{new IRF array}
\item{noise_L}{n x q matrix used to obtain tall IRF}
}
\description{
\code{lr_approx} calculates a tall IRF from a square
state space model using low-rank approximation
}
\keyword{internal}
