% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algo.R
\name{max_step}
\alias{max_step}
\title{The maximization step associated with the EM algorithm}
\usage{
max_step(out_smooth, tmpl_ss)
}
\arguments{
\item{out_smooth}{the output of \code{out_smooth}}

\item{tmpl_ss}{model template}
}
\value{
a list of elements
\item{model_new}{a state space model corresponding to the new estimated parameters}
\item{params_new}{a vector of deep parameters}
}
\description{
For details see section 3.2. in \emph{Estimation of Impulse-Response
Functions with Dynamic Factor Models: A New Parametrization}
available at \url{https://arxiv.org/pdf/2202.00310.pdf}.
}
\keyword{internal}
