% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intial_estimation.R
\name{normalize_irf}
\alias{normalize_irf}
\title{Normalize zero-lag coefficient of the IRF}
\usage{
normalize_irf(pseries_obj, type = c("identity", "lq_decomp"))
}
\arguments{
\item{pseries_obj}{IRF object, should have sufficiently many lags}

\item{type}{Character string. Either \code{identity} or \code{lq_decomp}.
In the former case the first submatrix of the tall transfer function will be transformed to the identity matrix.
In the latter case, this submatrix will be made lower triangular by an orthogonal transformation.}
}
\value{
a list of elements
\item{pseries_array}{the normalised transfer function as array}
\item{noise_L}{the inverse of the matrix used for normalising}
}
\description{
Depending on the template/identification scheme, the first \code{dim_in} rows of
the zero-lag coefficient of the transfer function are either normalized to the
identity matrix or lower triangular.
}
\keyword{internal}
