% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{pap_factory}
\alias{pap_factory}
\title{Append file name to path}
\usage{
pap_factory(path)
}
\arguments{
\item{path}{file path to append to \code{path}}
}
\value{
New file path
}
\description{
Path is in enclosing environment!
Helper for running scripts.
}
\examples{
path = "../local_data_ukko2/"
pap = pap_factory(path)
pap("myfile.whatever")
}
\keyword{internal}
