% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfd4dfm-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs}{First argument of the function of the right-hand-side of the pipe operator.}

\item{rhs}{Function whose first argument is given by the left-hand-side argument \code{lhs} of the pipe operator.}
}
\description{
Re-export pipe operator \code{\%>\%} to turn function composition into a series of imperative statements.
For more extensive description, see function \code{`\%>\%`} in package \emph{magrittr}.
}
