% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{prune_data}
\alias{prune_data}
\title{Obtain a balanced and possibly imputed and standardized panel of data}
\usage{
prune_data(
  df,
  int_vars = NULL,
  start_date = NULL,
  end_date = NULL,
  impute = TRUE,
  sdize = TRUE,
  trans_ix = NULL
)
}
\arguments{
\item{df}{\eqn{T x (N+1)} data panel with the first column specifying the time stamp,
\eqn{T} giving the number of observations and \eqn{N} the number of variables}

\item{int_vars}{vector of strings specifying the variables of interest, which are ordered first}

\item{start_date}{\link[lubridate]{ymd} object specifying the start date of the sample}

\item{end_date}{\link[lubridate]{ymd} object specifying the end date of the sample}

\item{impute}{Are the outliers removed and replaced with an imputed value
(for the imputation scheme see \link[fbi]{tw_apc})? Defaults to TRUE}

\item{sdize}{Are the variables standardized to have zero mean and unit standard deviation? Defaults to FALSE}

\item{trans_ix}{vector containing the transformation codes of the variables}
}
\value{
list of components
\item{df}{A \eqn{T x N} data frame corresponding transformed data}
\item{date}{vector of \link[lubridate]{ymd} components giving the sample span}
\item{discarded}{A list of discarded variables containing missing values}
\item{sd_mat}{If \code{sdize = TRUE}, a matrix containing standard deviations of the variables on the diagonal
is also supplied}
\item{trans_ix}{vector of reordered transformation codes if \code{int_vars} is supplied}
}
\description{
This convenience function allows the user to manipulate the supplied data matrix
with respect to
\itemize{
\item{the sample span;}
\item{reordering of the variables;}
\item{imputation of outliers; and}
\item{standardization.}
}
}
\keyword{internal}
