% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat.R
\name{pseries}
\alias{pseries}
\alias{pseries.default}
\alias{pseries.polm}
\alias{pseries.lmfd}
\alias{pseries.rmfd}
\alias{pseries.stsp}
\alias{pseries.lpolm}
\title{Power series Parameters}
\usage{
pseries(obj, lag.max, ...)

\method{pseries}{default}(obj, lag.max = 5, ...)

\method{pseries}{polm}(obj, lag.max = 5, ...)

\method{pseries}{lmfd}(obj, lag.max = 5, ...)

\method{pseries}{rmfd}(obj, lag.max = 5, ...)

\method{pseries}{stsp}(obj, lag.max = 5, ...)

\method{pseries}{lpolm}(obj, ...)
}
\arguments{
\item{obj}{(rational) matrix object, i.e. a \code{\link{polm}}, \code{\link{lmfd}},  \code{\link{rmfd}}, \code{\link{stsp}} object
or an object which may be coerced to a polynomial matrix with \code{polm(obj)}.
The default \code{S3} method first coerces the input argument \code{obj} to a \code{polm} object.
If this fails an error is thrown.}

\item{lag.max}{(integer) maximum lag.}

\item{...}{not used.}
}
\value{
object of type \code{pseries}
}
\description{
This function returns the coefficients of the power series expansion (around \eqn{z=0}) of
a rational matrix
\deqn{c(z) = c_0 + c_1 z + c_2 z^2 + \cdots}{c(z) = c[0] + c[1] z + c[2] z^2 + \dots}
Note that for a left matrix fraction description \eqn{c(z) = a(z)^{-1} b(z)} the
matrix \eqn{a(0)}  must be an invertible matrix.
If the rational functions corresponds to a VARMA or statespace model,
then this sequence of coefficients is also
called \emph{impulse response function}.
}
\details{
A \code{pseries} object is simply a 3-dimensional (numeric or complex) array
of dimension c(m,n,lag.max+1) with a class attribute c('pseries', 'ratm').
The dimensions \eqn{(m,n)} may also be zero.
}
\keyword{internal}
