% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_tools.R
\name{rmfd2stsp}
\alias{rmfd2stsp}
\title{Represent RMFD model in state space format}
\usage{
rmfd2stsp(rmfdsys, nu)
}
\arguments{
\item{rmfdsys}{\code{rmfd} object}

\item{nu}{vector of Kronecker indices}
}
\value{
\code{ABCD}, matrix of dimensions \eqn{r+n x r+n} of the corresponding state space model,
where state and output dimensions are denoted by \eqn{r} and \eqn{n}.
}
\description{
Represent RMFD model in state space format
}
\keyword{internal}
