% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algo.R
\name{smoothed_moments8}
\alias{smoothed_moments8}
\title{Obtaining Moment Matrices for EM Algorithm Using Kalman Filter/Smoother}
\usage{
smoothed_moments8(stsp_mod, Sigma, data_wide, only_ll = FALSE)
}
\arguments{
\item{stsp_mod}{\link{stspmod} object containing the state space model used for smoothing and the idiosyncratic noise component}

\item{Sigma}{Matrix of dimension \code{dim_in x dim_in}. All \code{dim_in^2} elements are saved, symmetry is not taken into account.}

\item{data_wide}{Matrix of dimension \code{dim_out x n_obs}}
}
\value{
List, see code
}
\description{
In contrast to the KFAS package, we take the covariance between the input terms in the state and observation equation into account.
Based on Hannan, Deistler (2012).
}
