% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat.R
\name{str methods}
\alias{str methods}
\alias{str.lpolm}
\alias{str.polm}
\alias{str.lmfd}
\alias{str.rmfd}
\alias{str.stsp}
\alias{str.pseries}
\alias{str.zvalues}
\title{Display the Structure of Objects}
\usage{
\method{str}{lpolm}(object, ...)

\method{str}{polm}(object, ...)

\method{str}{lmfd}(object, ...)

\method{str}{rmfd}(object, ...)

\method{str}{stsp}(object, ...)

\method{str}{pseries}(object, ...)

\method{str}{zvalues}(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{not used}
}
\value{
invisible(NULL)
}
\description{
Display the Structure of Objects
}
