% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_tools.R
\name{tmpl_rmfd_echelon_ss}
\alias{tmpl_rmfd_echelon_ss}
\title{Template for echelon RMFD model in state space format}
\usage{
tmpl_rmfd_echelon_ss(dim_out, nu, degs = NULL)
}
\arguments{
\item{dim_out}{output dimension, i.e. number of variables}

\item{nu}{vector of Kronecker indices}

\item{degs}{vector of length 2, where one can optionally specify the degrees of
\code{c(z)} and \code{d(z)}, either of the parameters must be equal to max(nu)}
}
\description{
In this template one obtains a pure echelon form by supplying the vector
of Kronecker indices and leaving the degrees of c(z) and d(z) unspecified.
On the other hand, the (P,Q)-form is obtained by supplying a vector
of Kronecker indices where all the values are equal to the maximum between
the degrees of c(z) and d(z), and then specifying the polynomial degrees
c(z) and d(z). Finally, a hybrid version of these two forms is obtained
by setting the Kronecker indices and polynomial degrees freely with
the restriction that \eqn{max(nu) = max(P, Q)}.
}
\keyword{internal}
