% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algo.R
\name{update_em2}
\alias{update_em2}
\title{Estimation of the D-DFM model using the EM algorithm}
\usage{
update_em2(
  params_deep_init,
  sigma_init,
  data_wide,
  tmpl_ss,
  MAXIT = 100,
  VERBOSE = FALSE,
  conv_crit = 0.01
)
}
\arguments{
\item{params_deep_init}{initial system and idiosyncratic variance parameter values}

\item{sigma_init}{initial value of the \eqn{dim_in x dim_in} innovation covariance matrix}

\item{data_wide}{\eqn{n x T} matrix containing the data}

\item{tmpl_ss}{\code{tmpl_rmfd_echelon_ss} object specifying the model structure}

\item{MAXIT}{integer denoting maximum number of iterations}

\item{VERBOSE}{boolean, print estimation progress?}
}
\description{
This function carries out the EM estimation of RMFD model by iterating between
E and M step (\code{smoothed_moments8} and \code{max_step}, resp.) until convergence
}
\keyword{internal}
