% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{boot_irfs}
\alias{boot_irfs}
\title{Obtain a bootstrap distribution of the IRFs}
\usage{
boot_irfs(
  df,
  nu,
  degs = NULL,
  maxit = 250,
  conv_crit = 0.001,
  verbose,
  h,
  nrep,
  int_vars,
  save_file = FALSE
)
}
\arguments{
\item{df}{\code{(T x N)}-dimensional data matrix}

\item{nu}{\code{q}-dimensional vector of Kronecker indices}

\item{degs}{optional, vector of length 2, defining the orders of c(z) and d(z)}

\item{maxit}{Maximum number of iterations in the estimation procedure, defaults to 250}

\item{conv_crit}{Convergence criterion in the estimation, defaults to 1e-3}

\item{verbose}{Should the estimation process be printed?}

\item{h}{the IRF horizon \eqn{h}}

\item{nrep}{number of bootstrap draws}

\item{int_vars}{\code{q}-dimensional vector coding the positions of the variables of interest}

\item{save_file}{save the IRF array at the end?}
}
\value{
An array of dimension \eqn{n x q h+1 x nrep} corresponding to the recursively identified
structural impulse-response function
}
\description{
Obtain a bootstrap distribution of the IRFs
}
\keyword{internal}
