% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{chol_ident}
\alias{chol_ident}
\title{Cholesky identification of the IRF}
\usage{
chol_ident(est_obj, int_vars, est_type = c("rmfd", "fglr"))
}
\arguments{
\item{est_obj}{\code{\link{estim_wrap}} or \code{\link{DfmRawImp}}
object containing the raw IRF and the matrix
used to obtain the lower triangular Cholesky factor}

\item{int_vars}{vector denoting the variables of interest and the respective ordering}

\item{est_type}{specify the DFM estimation method}
}
\value{
An IRF array of the same size as the raw IRF used as an input
}
\description{
\code{chol_ident} carries out the recursive identification using the Cholesky decomposition
of the \eqn{(q x q)} innovation covariance matrix. The resulting IRF has then
a structural interpretation depending on the ordering of the variables. The user
must specify whether the estimation method of the DFM is the S-DFM (\code{\link{DfmRawImp}})
or D-DFM (\code{\link{estim_wrap}}).
}
\keyword{internal}
