% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{DfmRawImp}
\alias{DfmRawImp}
\title{FGLR two-step method using VARs to estimate DFM}
\usage{
DfmRawImp(X, q, r, k, h)
}
\arguments{
\item{X}{Tall data matrix}

\item{q}{Number of dynamic factors,}

\item{r}{Number of static factors, must be larger than or equal to \eqn{q}}

\item{k}{Number of lags in the VAR for static factors}

\item{h}{number of lags included in the IRF}
}
\value{
The function \code{DfmRawImp} returns the following components
\item{var_model}{\code{\link{est_ar}} object corresponding to the VAR on \eqn{r} static factors}
\item{KM}{\eqn{r x q} matrix corresponding to the matrix used for low rank approximation
of the error term}
\item{WW}{\eqn{N}-dimensional diagonal matrix containing the standard deviations of the variables}
\item{W}{\eqn{N x r} factor loading matrix}
}
\description{
Calculate raw IRFs for S-DFM according to section 2.3.1 in
\emph{Estimation of Impulse-Response Functions with Dynamic Factor Models:
A New Parametrization} available at \url{https://arxiv.org/pdf/2202.00310.pdf}.
}
\examples{
X <- FRED_light$df
test <- DfmRawImp(X = X, q = 4, r = 16, k = 2, h = 50)
}
