% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_everything.R
\name{do_everything_rmfd}
\alias{do_everything_rmfd}
\title{Estimate IRFs and associated bootstrap intervals for D-DFM}
\usage{
do_everything_rmfd(
  df,
  r,
  h = 50,
  nrep = 0,
  conv_crit = 1e-05,
  ci = 0.8,
  init_rep = 0,
  verbose = FALSE,
  mod_str = NULL
)
}
\arguments{
\item{df}{list containing items:
\enumerate{
\item \code{df}, data in a tall matrix;
\item \code{int_ix}, index coding the rows of the interest variables in the data;
\item \code{trans_ix}, index giving the variable transformation codes
}}

\item{r}{static factor dimension}

\item{h}{h-step ahead IRFs}

\item{nrep}{number of replications in the bootstrap}

\item{conv_crit}{convergence criterion of the EM algorithm}

\item{ci}{confidence intervals for the IRFs}

\item{init_rep}{replications for the starting value procedure, for details, see \link[rmfd]{boot_init}}

\item{verbose}{logical, print the estimation process?}

\item{mod_str, }{list containing 1) the Kronecker index vector; 2) polynomial degrees of c(z) and d(z)}
}
\value{

}
\description{
Estimate IRFs and associated bootstrap intervals for D-DFM
}
