% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{irf_x}
\alias{irf_x}
\title{Left and right multiply IRFs with constant matrices}
\usage{
irf_x(irf, pre_mat = NULL, post_mat = NULL)
}
\arguments{
\item{irf}{\code{n x q x (h + 1)} array containing the impulse responses,
where \eqn{n, q} and \eqn{h} are the cross-sectional dimension, number of
dynamic factors, and IRF horizon, resepectively.}

\item{pre_mat}{matrix used in the left multiplication of the IRF}

\item{post_mat}{matrix used in the right multiplication of the IRF}
}
\value{
An array rotated in the third dimension.
}
\description{
\code{irf_x} is a helper function rotating arrays in the third
dimension of the given array.
}
\examples{
# create a random array corresponding to the impulse response function
irf_ex <- stats::rnorm(4*4*4) \%>\% array(dim = c(4,4,4))
# assume the recursive identification via cholesky is given as
chol_ex <- stats::rnorm(16) \%>\% matrix(4,4) \%>\% tcrossprod() \%>\% chol() \%>\% t()
# then the recursive identification of the IRF can be obtained as
(sirf_ex <- irf_x(irf_ex, post_mat = solve(irf_ex[,,1])\%*\%chol_ex) \%>\% zapsmall())

}
