% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intial_estimation.R
\name{lr_approx}
\alias{lr_approx}
\title{Get Tall IRFs}
\usage{
lr_approx(ssmod, dim_in, lag.max = 30)
}
\arguments{
\item{ssmod}{\link[rmfd2dfm]{stspmod} object of the square state space model.}

\item{dim_in}{the input dimension of the transformed IRF.}

\item{lag.max}{maximum lag to be passed onto function \link[rationalmatrices]{pseries}, default is 30.}
}
\value{
a list of elements:
\item{irf}{new IRF array}
\item{noise_L}{n x q matrix used to obtain tall IRF}
}
\description{
\code{lr_approx} This helper function calculates a tall IRF from a square
state space model using low-rank approximation.
}
\keyword{internal}
