% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{plot_irfs}
\alias{plot_irfs}
\title{Plot impulse response functions}
\usage{
plot_irfs(irf_arr, var_names, plot_title, label_y = TRUE)
}
\arguments{
\item{irf_arr}{Matrix or array of dimension \eqn{h+1 x q} or \eqn{(h+1 x 3 x q)},
where \eqn{h} is the number of periods in IRF and \eqn{q} is the number of
structural shocks}

\item{var_names}{vector variable names of the \eqn{q} impacted variables}

\item{plot_title}{Title of the figure}
}
\value{
list containing the figures
}
\description{
This function plots impulse response functions for a given array
containing the impulse responses and error bands. This function
returns a list of figures containing IRFs per variable.
}
