% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_tools.R
\name{stsp2rmfd}
\alias{stsp2rmfd}
\title{Obtain RMFD object from a state space system}
\usage{
stsp2rmfd(stspsys, dim_out, nu, degs = NULL)
}
\arguments{
\item{stspsys}{matrix corresponding to ABCD state space system}

\item{dim_out}{output dimension, i.e. number of variables}

\item{nu}{vector of right Kronecker indices}

\item{degs}{vector of length 2, where one can optionally specify the degrees of c(z) and d(z), either of the parameters must be equal to max(nu)}

\item{addnoise}{is the model augmented with idiosyncratic noise component?}
}
\value{
\code{rmfd} object corresponding to the state space setup
}
\description{
Obtain RMFD object from a state space system
}
\keyword{internal}
