% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{X_boot}
\alias{X_boot}
\title{Obtain a block bootstrap data sample}
\usage{
X_boot(X, L)
}
\arguments{
\item{X}{data matrix}

\item{L}{determines the block size}
}
\value{
bootstrapped data matrix of the same dimensions as the original one
}
\description{
\code{X_boot} performs a simple block bootstrap for the data matrix.
This is a direct copy of the code used in Forni and Gambetti (2010)
available at \url{http://pareto.uab.es/lgambetti/ReplicaForniGambettiJME.zip}.
}
\keyword{internal}
