% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_tools.R
\name{X_boot}
\alias{X_boot}
\title{NO EXPORT:: Obtain a block bootstrap sample}
\usage{
X_boot(X, L)
}
\arguments{
\item{X}{data matrix}

\item{L}{determines the size of block}
}
\value{

}
\description{
This function performs a simple block bootstrap for the data matrix.
This is a direct copy of FG code used in Forni and Gambetti (2010)
}
\examples{
rnorm(100)
}
